deduplicate Sys_Print()
