+++ sys_sdl.c
@@ -1,13 +1,7 @@
 #ifdef WIN32
-#include <io.h> // Include this BEFORE darkplaces.h because it uses strncpy which trips DP_STATIC_ASSERT
 #else
-#include <unistd.h>
 #include <fcntl.h>
 #include <sys/time.h>
-#endif
-
-#ifdef __ANDROID__
-#include <android/log.h>
 #endif
 
 /*
@@ -50,41 +44,6 @@ void Sys_SDL_Dialog(const char *title, const char *string)
 {
 	if(!nocrashdialog)
 		SDL_ShowSimpleMessageBox(SDL_MESSAGEBOX_ERROR, title, string, NULL);
-}
-
-void Sys_Print(const char *text)
-{
-#ifdef __ANDROID__
-	if (developer.integer > 0)
-	{
-		__android_log_write(ANDROID_LOG_DEBUG, sys.argv[0], text);
-	}
-#else
-	if(sys.outfd < 0)
-		return;
-#ifndef WIN32
-	// BUG: for some reason, NDELAY also affects stdout (1) when used on stdin (0).
-	// this is because both go to /dev/tty by default!
-	{
-		int origflags = fcntl (sys.outfd, F_GETFL, 0);
-		fcntl (sys.outfd, F_SETFL, origflags & ~O_NONBLOCK);
-#endif
-#ifdef WIN32
-#define write _write
-#endif
-		while(*text)
-		{
-			fs_offset_t written = (fs_offset_t)write(sys.outfd, text, (int)strlen(text));
-			if(written <= 0)
-				break; // sorry, I cannot do anything about this error - without an output
-			text += written;
-		}
-#ifndef WIN32
-		fcntl (sys.outfd, F_SETFL, origflags);
-	}
-#endif
-	//fprintf(stdout, "%s", text);
-#endif
 }
 
 char *Sys_GetClipboardData (void)
