+++ sys_unix.c
@@ -2,10 +2,8 @@
 #ifdef WIN32
 #include <windows.h>
 #include <mmsystem.h>
-#include <io.h>
 #else
 #include <sys/time.h>
-#include <unistd.h>
 #include <fcntl.h>
 #endif
 
@@ -26,33 +24,6 @@ void Sys_Shutdown (void)
 
 void Sys_SDL_Dialog(const char *title, const char *string)
 {
-}
-
-void Sys_Print(const char *text)
-{
-	if(sys.outfd < 0)
-		return;
-	// BUG: for some reason, NDELAY also affects stdout (1) when used on stdin (0).
-	// this is because both go to /dev/tty by default!
-	{
-#ifndef WIN32
-		int origflags = fcntl (sys.outfd, F_GETFL, 0);
-		fcntl (sys.outfd, F_SETFL, origflags & ~O_NONBLOCK);
-#else
-#define write _write
-#endif
-		while(*text)
-		{
-			fs_offset_t written = (fs_offset_t)write(sys.outfd, text, (int)strlen(text));
-			if(written <= 0)
-				break; // sorry, I cannot do anything about this error - without an output
-			text += written;
-		}
-#ifndef WIN32
-		fcntl (sys.outfd, F_SETFL, origflags);
-#endif
-	}
-	//fprintf(stdout, "%s", text);
 }
 
 char *Sys_GetClipboardData (void)
